feat: misc debug was updated
