--- misc/to_string_other.cpp
@@ -1,23 +1,0 @@
-template <typename A, typename B>
-string to_string(const pair<A, B> p) {
-    return "(" + to_string(p.first) + ", " + to_string(p.second) + ")";
-}
-
-template <typename A, typename B, typename C>
-string to_string(const tuple<A, B, C> p) {
-    return "(" + to_string(get<0>(p)) + ", " + to_string(get<1>(p)) + ", " + to_string(get<2>(p)) + ")";
-}
- 
-template <typename A, typename B, typename C, typename D>
-string to_string(const tuple<A, B, C, D> p) {
-    return "(" + to_string(get<0>(p)) + ", " + to_string(get<1>(p)) + ", " + to_string(get<2>(p)) + ", " + to_string(get<3>(p)) + ")";
-}
-
-template <size_t N>
-string to_string(const bitset<N> v) {
-    string res = "";
-    for (size_t i = 0; i < N; i++) {
-        res += static_cast<char>('0' + v[i]);
-    }
-    return res;
-}
